feature: Implement parallel task group
The conquer part is parallelized. One big issue is that the algorithms
are numerically unstable. A dense set of points might not result in a
convex hull because of floating point errors. On the other hand, we
need a very large (therefore dense) input to see the speedup of the
concurrent algorithm in action.